more tests around the boosted fees

+++ src/interfaces/extensions/IBoostedFees.sol
@@ -12,12 +12,12 @@ import {PoolId} from "../../types/poolId.sol";
 /// @dev Configuration occurs via forwarded calls; hooks trigger accrual by locking the Core contract and accumulating fees.
 interface IBoostedFees is IExposedStorage, IExtension, ILocker, IForwardee {
     /// @notice Emitted when a pool has incentives added
-    /// @param poolId Pool identifier
-    /// @param startTime Incentive start time (inclusive)
-    /// @param endTime Incentive end time (exclusive)
-    /// @param rate0 Per-second incentive rate for token0 as a fixed point *.32 value
-    /// @param rate1 Per-second incentive rate for token1 as a fixed point *.32 value
-    event PoolBoosted(PoolId indexed poolId, uint64 startTime, uint64 endTime, uint256 rate0, uint256 rate1);
+    /// @param poolId Identifier of the pool that is being incentivized
+    /// @param startTime The first second that incentives will be paid out
+    /// @param endTime The time at which incentives stop paying out
+    /// @param rate0 Per-second incentive rate for token0 as a fixed point 96.32 value
+    /// @param rate1 Per-second incentive rate for token1 as a fixed point 96.32 value
+    event PoolBoosted(PoolId indexed poolId, uint64 startTime, uint64 endTime, uint128 rate0, uint128 rate1);
 
     /// @notice Reverts when provided times are invalid.
     error InvalidTimestamps();
